Move useProjectTask to correct folder
