+++ ui/src/hooks/use-project-task.hook.ts
@@ -1,7 +1,7 @@
 // Copyright (C) 2025 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
-import { $api } from '../../../api/client';
+import { $api } from '../api/client';
 import { useProjectIdentifier } from './use-project-identifier.hook';
 
 export const useProjectTask = () => {
